Added additional fields to SITL parsing, example query, TODOs
Changes to parsing for adding columns to SITL dataframe, converted FOM to FLOAT, and added a boolean for searching the Discussion field. Presented an example query to identify high quality BBFs from Discussion text. Still need to add a duration field, calculated from datetime. Added a note, since we could have a manual feedback leading to an annotation in the SITL dataframe for confirmation of BBF for ML...