Added new parsing code to create calculated fields which can be queried to I.D. BBF/DF start/end time events
Needs work, but this is a start! I need to add a duration of event parameter and also if possible have a way to parse MULTI events -- BBFs and DFs in the same time range. I think we can use the time range as the foriegn key between the SITL reports and the measurement data. This *hopefully* can be ingested into the ML code Rachel was working on. Thanks, Jack